+++ src/sagemaker/processing.py
@@ -38,15 +38,8 @@ from sagemaker.workflow.parameters import Parameter
 from sagemaker.workflow.entities import Expression
 from sagemaker.dataset_definition.inputs import S3Input, DatasetDefinition
 from sagemaker.apiutils._base_types import ApiObject
-
-##
 from sagemaker.estimator import Framework
-from sagemaker.mxnet.estimator import MXNet
-from sagemaker.pytorch.estimator import PyTorch
 from sagemaker.s3 import S3Uploader
-from sagemaker.sklearn.estimator import SKLearn
-from sagemaker.tensorflow.estimator import TensorFlow
-from sagemaker.xgboost.estimator import XGBoost
 
 
 logger = logging.getLogger(__name__)
@@ -1548,296 +1541,3 @@ python {entry_point} "$@"
             ProcessingInput(source=s3_payload, destination="/opt/ml/processing/input/code/payload/")
         )
         return inputs
-
-
-class MXNetProcessor(FrameworkProcessor):
-    """Handles Amazon SageMaker processing tasks for jobs using MXNet containers."""
-
-    estimator_cls = MXNet
-
-    def __init__(
-        self,
-        framework_version: str,  # New arg
-        s3_prefix: str,  # New arg
-        role: str,
-        instance_count: int,
-        instance_type: str,
-        py_version: str = "py3",  # New kwarg
-        image_uri: Optional[str] = None,
-        volume_size_in_gb: int = 30,
-        volume_kms_key: Optional[str] = None,
-        output_kms_key: Optional[str] = None,
-        max_runtime_in_seconds: Optional[int] = None,
-        base_job_name: Optional[str] = None,
-        sagemaker_session: Optional[Session] = None,
-        env: Optional[Dict[str, str]] = None,
-        tags: Optional[List[Dict[str, Any]]] = None,
-        network_config: Optional[NetworkConfig] = None,
-    ):
-        """This processor executes a Python script in a managed MXNet execution environment.
-
-        Unless ``image_uri`` is specified, the MXNet environment is an
-        Amazon-built Docker container that executes functions defined in the supplied
-        ``entry_point`` Python script.
-
-        The arguments have the exact same meaning as in ``FrameworkProcessor``.
-
-        .. tip::
-
-            You can find additional parameters for initializing this class at
-            :class:`~smallmatter.ds.FrameworkProcessor`.
-        """
-        super().__init__(
-            self.estimator_cls,
-            framework_version,
-            s3_prefix,
-            role,
-            instance_count,
-            instance_type,
-            py_version,
-            image_uri,
-            volume_size_in_gb,
-            volume_kms_key,
-            output_kms_key,
-            max_runtime_in_seconds,
-            base_job_name,
-            sagemaker_session,
-            env,
-            tags,
-            network_config,
-        )
-
-
-class PyTorchProcessor(FrameworkProcessor):
-    """Handles Amazon SageMaker processing tasks for jobs using PyTorch containers."""
-
-    estimator_cls = PyTorch
-
-    def __init__(
-        self,
-        framework_version: str,  # New arg
-        s3_prefix: str,  # New arg
-        role: str,
-        instance_count: int,
-        instance_type: str,
-        py_version: str = "py3",  # New kwarg
-        image_uri: Optional[str] = None,
-        volume_size_in_gb: int = 30,
-        volume_kms_key: Optional[str] = None,
-        output_kms_key: Optional[str] = None,
-        max_runtime_in_seconds: Optional[int] = None,
-        base_job_name: Optional[str] = None,
-        sagemaker_session: Optional[Session] = None,
-        env: Optional[Dict[str, str]] = None,
-        tags: Optional[List[Dict[str, Any]]] = None,
-        network_config: Optional[NetworkConfig] = None,
-    ):
-        """This processor executes a Python script in a PyTorch execution environment.
-
-        Unless ``image_uri`` is specified, the PyTorch environment is an
-        Amazon-built Docker container that executes functions defined in the supplied
-        ``entry_point`` Python script.
-
-        The arguments have the exact same meaning as in ``FrameworkProcessor``.
-
-        .. tip::
-
-            You can find additional parameters for initializing this class at
-            :class:`~smallmatter.ds.FrameworkProcessor`.
-        """
-        super().__init__(
-            self.estimator_cls,
-            framework_version,
-            s3_prefix,
-            role,
-            instance_count,
-            instance_type,
-            py_version,
-            image_uri,
-            volume_size_in_gb,
-            volume_kms_key,
-            output_kms_key,
-            max_runtime_in_seconds,
-            base_job_name,
-            sagemaker_session,
-            env,
-            tags,
-            network_config,
-        )
-
-
-class SKLearnProcessorAlt(FrameworkProcessor):
-    """Handles Amazon SageMaker processing tasks for jobs using scikit-learn containers."""
-
-    estimator_cls = SKLearn
-
-    def __init__(
-        self,
-        framework_version: str,  # New arg
-        s3_prefix: str,  # New arg
-        role: str,
-        instance_count: int,
-        instance_type: str,
-        py_version: str = "py3",  # New kwarg
-        image_uri: Optional[str] = None,
-        volume_size_in_gb: int = 30,
-        volume_kms_key: Optional[str] = None,
-        output_kms_key: Optional[str] = None,
-        max_runtime_in_seconds: Optional[int] = None,
-        base_job_name: Optional[str] = None,
-        sagemaker_session: Optional[Session] = None,
-        env: Optional[Dict[str, str]] = None,
-        tags: Optional[List[Dict[str, Any]]] = None,
-        network_config: Optional[NetworkConfig] = None,
-    ):
-        """This processor executes a Python script in a scikit-learn execution environment.
-
-        This class has an 'Alt' suffix to denote it as an alternative to built-in
-        ``sagemaker.sklearn.processing.SKLearnProcessor``.
-
-        Unless ``image_uri`` is specified, the scikit-learn environment is an
-        Amazon-built Docker container that executes functions defined in the supplied
-        ``entry_point`` Python script.
-
-        The arguments have the exact same meaning as in ``FrameworkProcessor``.
-
-        .. tip::
-
-            You can find additional parameters for initializing this class at
-            :class:`~smallmatter.ds.FrameworkProcessor`.
-        """
-        super().__init__(
-            self.estimator_cls,
-            framework_version,
-            s3_prefix,
-            role,
-            instance_count,
-            instance_type,
-            py_version,
-            image_uri,
-            volume_size_in_gb,
-            volume_kms_key,
-            output_kms_key,
-            max_runtime_in_seconds,
-            base_job_name,
-            sagemaker_session,
-            env,
-            tags,
-            network_config,
-        )
-
-
-class TensorFlowProcessor(FrameworkProcessor):
-    """Handles Amazon SageMaker processing tasks for jobs using TensorFlow containers."""
-
-    estimator_cls = TensorFlow
-
-    def __init__(
-        self,
-        framework_version: str,  # New arg
-        s3_prefix: str,  # New arg
-        role: str,
-        instance_count: int,
-        instance_type: str,
-        py_version: str = "py3",  # New kwarg
-        image_uri: Optional[str] = None,
-        volume_size_in_gb: int = 30,
-        volume_kms_key: Optional[str] = None,
-        output_kms_key: Optional[str] = None,
-        max_runtime_in_seconds: Optional[int] = None,
-        base_job_name: Optional[str] = None,
-        sagemaker_session: Optional[Session] = None,
-        env: Optional[Dict[str, str]] = None,
-        tags: Optional[List[Dict[str, Any]]] = None,
-        network_config: Optional[NetworkConfig] = None,
-    ):
-        """This processor executes a Python script in a TensorFlow execution environment.
-
-        Unless ``image_uri`` is specified, the TensorFlow environment is an
-        Amazon-built Docker container that executes functions defined in the supplied
-        ``entry_point`` Python script.
-
-        The arguments have the exact same meaning as in ``FrameworkProcessor``.
-
-        .. tip::
-
-            You can find additional parameters for initializing this class at
-            :class:`~smallmatter.ds.FrameworkProcessor`.
-        """
-        super().__init__(
-            self.estimator_cls,
-            framework_version,
-            s3_prefix,
-            role,
-            instance_count,
-            instance_type,
-            py_version,
-            image_uri,
-            volume_size_in_gb,
-            volume_kms_key,
-            output_kms_key,
-            max_runtime_in_seconds,
-            base_job_name,
-            sagemaker_session,
-            env,
-            tags,
-            network_config,
-        )
-
-
-class XGBoostEstimator(FrameworkProcessor):
-    """Handles Amazon SageMaker processing tasks for jobs using XGBoost containers."""
-
-    estimator_cls = XGBoost
-
-    def __init__(
-        self,
-        framework_version: str,  # New arg
-        s3_prefix: str,  # New arg
-        role: str,
-        instance_count: int,
-        instance_type: str,
-        py_version: str = "py3",  # New kwarg
-        image_uri: Optional[str] = None,
-        volume_size_in_gb: int = 30,
-        volume_kms_key: Optional[str] = None,
-        output_kms_key: Optional[str] = None,
-        max_runtime_in_seconds: Optional[int] = None,
-        base_job_name: Optional[str] = None,
-        sagemaker_session: Optional[Session] = None,
-        env: Optional[Dict[str, str]] = None,
-        tags: Optional[List[Dict[str, Any]]] = None,
-        network_config: Optional[NetworkConfig] = None,
-    ):
-        """This processor executes a Python script in an XGBoost execution environment.
-
-        Unless ``image_uri`` is specified, the XGBoost environment is an Amazon-built
-        Docker container that executes functions defined in the supplied ``entry_point``
-        Python script.
-
-        The arguments have the exact same meaning as in ``FrameworkProcessor``.
-
-        .. tip::
-
-            You can find additional parameters for initializing this class at
-            :class:`~smallmatter.ds.FrameworkProcessor`.
-        """
-        super().__init__(
-            self.estimator_cls,
-            framework_version,
-            s3_prefix,
-            role,
-            instance_count,
-            instance_type,
-            py_version,
-            image_uri,
-            volume_size_in_gb,
-            volume_kms_key,
-            output_kms_key,
-            max_runtime_in_seconds,
-            base_job_name,
-            sagemaker_session,
-            env,
-            tags,
-            network_config,
-        )
